+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/analytics/AnalyticsListener.java
@@ -22,6 +22,7 @@ import static java.lang.annotation.ElementType.METHOD;
 import static java.lang.annotation.ElementType.PARAMETER;
 import static java.lang.annotation.ElementType.TYPE_USE;
 
+import android.media.AudioTrack;
 import android.media.MediaCodec;
 import android.media.MediaCodec.CodecException;
 import android.os.Looper;
@@ -232,6 +233,8 @@ public interface AnalyticsListener {
     EVENT_PLAYER_RELEASED,
     EVENT_AUDIO_CODEC_ERROR,
     EVENT_VIDEO_CODEC_ERROR,
+    EVENT_AUDIO_TRACK_INITIALIZED,
+    EVENT_AUDIO_TRACK_RELEASED
   })
   @interface EventFlags {}
 
@@ -434,6 +437,12 @@ public interface AnalyticsListener {
 
   /** The video codec encountered an error. */
   @UnstableApi int EVENT_VIDEO_CODEC_ERROR = 1030;
+
+  /** An audio track has been initialized. */
+  @UnstableApi int EVENT_AUDIO_TRACK_INITIALIZED = 1031;
+
+  /** An audio track has been released. */
+  @UnstableApi int EVENT_AUDIO_TRACK_RELEASED = 1032;
 
   /** Time information of an event. */
   @UnstableApi
@@ -1112,6 +1121,28 @@ public interface AnalyticsListener {
    */
   @UnstableApi
   default void onAudioCodecError(EventTime eventTime, Exception audioCodecError) {}
+
+  /**
+   * Called when an {@link AudioTrack} has been initialized.
+   *
+   * @param eventTime The event time.
+   * @param audioTrackConfig The {@link AudioSink.AudioTrackConfig} of the initialized {@link
+   *     AudioTrack}.
+   */
+  @UnstableApi
+  default void onAudioTrackInitialized(
+      EventTime eventTime, AudioSink.AudioTrackConfig audioTrackConfig) {}
+
+  /**
+   * Called when an {@link AudioTrack} has been released.
+   *
+   * @param eventTime The event time.
+   * @param audioTrackConfig The {@link AudioSink.AudioTrackConfig} of the released {@link
+   *     AudioTrack}.
+   */
+  @UnstableApi
+  default void onAudioTrackReleased(
+      EventTime eventTime, AudioSink.AudioTrackConfig audioTrackConfig) {}
 
   /**
    * Called when the volume changes.