+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/util/EventLogger.java
@@ -38,6 +38,7 @@ import androidx.media3.common.util.UnstableApi;
 import androidx.media3.exoplayer.DecoderCounters;
 import androidx.media3.exoplayer.DecoderReuseEvaluation;
 import androidx.media3.exoplayer.analytics.AnalyticsListener;
+import androidx.media3.exoplayer.audio.AudioSink;
 import androidx.media3.exoplayer.drm.DrmSession;
 import androidx.media3.exoplayer.source.LoadEventInfo;
 import androidx.media3.exoplayer.source.MediaLoadData;
@@ -403,6 +404,20 @@ public class EventLogger implements AnalyticsListener {
 
   @UnstableApi
   @Override
+  public void onAudioTrackInitialized(
+      EventTime eventTime, AudioSink.AudioTrackConfig audioTrackConfig) {
+    logd(eventTime, "audioTrackInit", getAudioTrackConfigString(audioTrackConfig));
+  }
+
+  @UnstableApi
+  @Override
+  public void onAudioTrackReleased(
+      EventTime eventTime, AudioSink.AudioTrackConfig audioTrackConfig) {
+    logd(eventTime, "audioTrackReleased", getAudioTrackConfigString(audioTrackConfig));
+  }
+
+  @UnstableApi
+  @Override
   public void onVideoEnabled(EventTime eventTime, DecoderCounters decoderCounters) {
     logd(eventTime, "videoEnabled");
   }
@@ -744,5 +759,19 @@ public class EventLogger implements AnalyticsListener {
       default:
         return "?";
     }
+  }
+
+  private static String getAudioTrackConfigString(AudioSink.AudioTrackConfig audioTrackConfig) {
+    return audioTrackConfig.encoding
+        + ","
+        + audioTrackConfig.channelConfig
+        + ","
+        + audioTrackConfig.sampleRate
+        + ","
+        + audioTrackConfig.tunneling
+        + ","
+        + audioTrackConfig.offload
+        + ","
+        + audioTrackConfig.bufferSize;
   }
 }